Protecting KeyMaterial with Keychain.
- Quick note about what KeyMaterial is / contains :

KeyMaterial is a class that abstract Cryptographic Keys management to talk to Trustline-token.
During Pairing, Trustline-Token generates 3 Keys : PassKey, CrKey and ComKey (see Wiki for details) and sends them to Trustline-Controller.
Trustline-Controller then:
  *generates a QRCode that contains theses 3 keys: your job is to print this QrCode, and to keep it somewhere safe.
  *saves only TWO keys : CrKey (for authentication / anti-replay) and ComKey (kind of pre-shared key to cipher communications).

Note that PassKey is NEVER stored anywhere, but inside Trustline-Token itself.

On Trustline-Controller, CrKey and ReqKey are saved using KeyChain properties : "WhenUnlockThisDeviceOnly" which means:
"Item data can only be accessed while the device is unlocked. This is recommended for items that only need be accessible while the application is in the foreground. Items with this attribute will never migrate to a new device, so after a backup is restored to a new device, these items will be missing."

What this means:
- Data is encrypted when the phone is locked this a unique key that's inside iPhone chip only.
- No, no backup is made during iTunes Synchro, cause you won't be able to restore this encrypted data on another Phone/Tablet (because chips will differ, and they won't have the same secret key).

Note that is theses key get compromised, your passwords will be safe as long as the attacker cannot access your Trustline Token that contains the PassKey.

So, How to use my iCloud synchronized encrypted password across multiple devices?

If you paired Trustline-Token with your phone, and you want to use another Phone / Tablet:
 - install Trustline-Controller on the over device.
 - make sure your iCloud account is activated on both devices.
 - instead of pairing, use the "pair with existing Trustline-Token" button and place your secret QRCode that contains the entire KeyMaterial in front of the Camera

Trustline-Controller will then import KeyMaterial CrKey and ComKey, so it can discuss and authenticate itself with the already-paired trustline-Token, and ask it to decrypt your synchronized encrypted password!